moves FormFieldBase type to payload
