+++ docs/migration-guide/overview.mdx
@@ -335,7 +335,6 @@ import {
   FieldMap,
   File,
   Form,
-  FormFieldBase,
   FormLoadingOverlayToggle,
   FormSubmit,
   GenerateConfirmation,
