+++ packages/richtext-lexical/src/features/blocks/component/BlockContent.tsx
@@ -1,6 +1,5 @@
-import type { FormFieldBase } from '@payloadcms/ui'
 import type { FieldMap } from '@payloadcms/ui/utilities/buildComponentMap'
-import type { CollapsedPreferences, Data, FormState } from 'payload'
+import type { CollapsedPreferences, Data, FormFieldBase, FormState } from 'payload'
 
 import { useLexicalComposerContext } from '@lexical/react/LexicalComposerContext.js'
 import { getTranslation } from '@payloadcms/translations'
