+++ packages/richtext-slate/src/field/elements/relationship/Element/index.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import type { FormFieldBase } from '@payloadcms/ui'
+import type { FormFieldBase } from 'payload'
 
 import { getTranslation } from '@payloadcms/translations'
 import {
