+++ packages/richtext-slate/src/field/providers/LeafButtonProvider.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import type { FormFieldBase } from '@payloadcms/ui'
+import type { FormFieldBase } from 'payload'
 
 import React from 'react'
