+++ packages/ui/src/fields/Email/index.tsx
@@ -4,7 +4,7 @@ import type { ClientValidate, EmailField as EmailFieldType } from 'payload'
 import { getTranslation } from '@payloadcms/translations'
 import React, { useCallback } from 'react'
 
-import type { FormFieldBase } from '../shared/index.js'
+import type { FormFieldBase } from 'payload'
 
 import { useFieldProps } from '../../forms/FieldPropsProvider/index.js'
 import { useField } from '../../forms/useField/index.js'
