+++ packages/ui/src/fields/Point/index.tsx
@@ -12,7 +12,7 @@ import './index.scss'
 
 const baseClass = 'point'
 
-import type { FormFieldBase } from '../shared/index.js'
+import type { FormFieldBase } from 'payload'
 
 import { useFieldProps } from '../../forms/FieldPropsProvider/index.js'
 import { FieldDescription } from '../FieldDescription/index.js'
