+++ packages/ui/src/fields/RadioGroup/index.tsx
@@ -14,7 +14,7 @@ import './index.scss'
 
 const baseClass = 'radio-group'
 
-import type { FormFieldBase } from '../shared/index.js'
+import type { FormFieldBase } from 'payload'
 
 import { useFieldProps } from '../../forms/FieldPropsProvider/index.js'
 import { FieldDescription } from '../FieldDescription/index.js'
