+++ packages/ui/src/fields/RichText/index.tsx
@@ -1,7 +1,7 @@
 import type React from 'react'
 
 import type { MappedField } from '../../providers/ComponentMap/buildComponentMap/types.js'
-import type { FormFieldBase } from '../shared/index.js'
+import type { FormFieldBase } from 'payload'
 
 export type RichTextFieldProps = {
   name: string
