+++ packages/ui/src/fields/Textarea/types.ts
@@ -2,7 +2,7 @@ import type { TextareaField as TextareaFieldType } from 'payload'
 
 import { type ChangeEvent } from 'react'
 
-import type { FormFieldBase } from '../shared/index.js'
+import type { FormFieldBase } from 'payload'
 
 export type TextareaFieldProps = {
   maxLength?: number
