+++ packages/ui/src/fields/shared/index.tsx
@@ -1,41 +1,6 @@
-import type {
-  DocumentPreferences,
-  ErrorProps,
-  FieldDescriptionProps,
-  LabelStatic,
-  Locale,
-  SanitizedLabelProps,
-  SanitizedLocalizationConfig,
-  User,
-  Validate,
-} from 'payload'
+import type { Locale, SanitizedLocalizationConfig } from 'payload'
 
 export const fieldBaseClass = 'field-type'
-
-export type FormFieldBase = {
-  AfterInput?: React.ReactNode
-  BeforeInput?: React.ReactNode
-  CustomDescription?: React.ReactNode
-  CustomError?: React.ReactNode
-  CustomLabel?: React.ReactNode
-  className?: string
-  custom?: Record<string, any>
-  descriptionProps?: FieldDescriptionProps
-  disabled?: boolean
-  docPreferences?: DocumentPreferences
-  errorProps?: ErrorProps
-  label?: LabelStatic
-  labelProps?: SanitizedLabelProps
-  locale?: Locale
-  localized?: boolean
-  path?: string
-  readOnly?: boolean
-  required?: boolean
-  rtl?: boolean
-  style?: React.CSSProperties
-  user?: User
-  validate?: Validate
-}
 
 /**
  * Determines whether a field should be displayed as right-to-left (RTL) based on its configuration, payload's localization configuration and the adming user's currently enabled locale.
